Add features : Auth & add file as pic

+++ backend/controllers/stuff.js
@@ -1,11 +1,14 @@
 const Thing = require('../models/Thing');
+const fs = require('fs');
 
 // --- POST ---
 // POST one stuff
 exports.createThing = (req, res, next) => {
-  delete req.body._id;
+  const thingObject = JSON.parse(req.body.thing);
+  delete thingObject._id;
   const thing = new Thing({
-    ...req.body
+    ...thingObject,
+    imageUrl: `${req.protocol}://${req.get('host')}/images/${req.file.filename}`
   });
   thing
     .save()
@@ -31,7 +34,13 @@ exports.getOneThing = (req, res, next) => {
 // --- PUT ---
 // PUT One Stuff (Update)
 exports.modifyThing = (req, res, next) => {
-  Thing.updateOne({ _id: req.params.id }, { ...req.body, _id: req.params.id })
+  const thingObject = req.file
+    ? {
+        ...JSON.parse(req.body.thing),
+        imageUrl: `${req.protocol}://${req.get('host')}/images/${req.file.filename}`
+      }
+    : { ...req.body };
+  Thing.updateOne({ _id: req.params.id }, { ...thingObject, _id: req.params.id })
     .then(() => res.status(200).json({ message: 'Objet modifié avec succés' }))
     .catch(error => res.status(400).json(error));
 };
@@ -39,7 +48,24 @@ exports.modifyThing = (req, res, next) => {
 // --- DELETE ---
 // DELETE One Stuff
 exports.deleteThing = (req, res, next) => {
-  Thing.deleteOne({ _id: req.params.id })
-    .then(() => res.status(200).json({ message: 'Objet supprimé avec succés' }))
-    .catch(error => res.status(400).json(error));
+  Thing.findOne({ _id: req.params.id })
+    .then(thing => {
+      if (!thing) {
+        return res.status(404).json({
+          error: new Error('Objet non trouvé !')
+        });
+      }
+      if (thing.userId !== req.auth.userId) {
+        return res.status(401).json({
+          error: new Error('Requête non autorisée')
+        });
+      }
+      const filename = thing.imageUrl.split('/images/')[1];
+      fs.unlink(`images/${filename}`, () => {
+        Thing.deleteOne({ _id: req.params.id })
+          .then(() => res.status(200).json({ message: 'Objet supprimé avec succés' }))
+          .catch(error => res.status(400).json(error));
+      });
+    })
+    .catch(error => res.status(500).json({ error }));
 };
